+++ Code/Engine/Foundation/Basics/AllDefinesOff.h
@@ -70,6 +70,7 @@
 #define EZ_SUPPORTS_CASE_INSENSITIVE_PATHS EZ_OFF
 #define EZ_SUPPORTS_CRASH_DUMPS EZ_OFF
 #define EZ_SUPPORTS_LONG_PATHS EZ_OFF
+#define EZ_SUPPORTS_IPC EZ_OFF
 
 // Allocators
 #define EZ_ALLOC_GUARD_ALLOCATIONS EZ_OFF
