+++ Code/Engine/Foundation/Platform/Linux/Features_Platform.h
@@ -74,6 +74,10 @@
 #undef EZ_SUPPORTS_PROCESSES
 #define EZ_SUPPORTS_PROCESSES EZ_ON
 
+/// Whether inter-process communication via pipes is supported
+#undef EZ_SUPPORTS_IPC
+#define EZ_SUPPORTS_IPC EZ_ON
+
 // SIMD support
 #undef EZ_SIMD_IMPLEMENTATION
