+++ Code/Engine/Foundation/Platform/Win/Features_Platform.h
@@ -70,6 +70,10 @@
 #undef EZ_SUPPORTS_PROCESSES
 #define EZ_SUPPORTS_PROCESSES EZ_ON
 
+/// Whether inter-process communication via pipes is supported
+#undef EZ_SUPPORTS_IPC
+#define EZ_SUPPORTS_IPC EZ_ON
+
 // SIMD support
 #undef EZ_SIMD_IMPLEMENTATION
