+++ Code/Engine/Foundation/Platform/Win/PipeChannel_Platform.h
@@ -66,4 +66,6 @@ private:
   ezUInt8 m_InputBuffer[BUFFER_SIZE];
 };
 
+using ezPipeChannel_Platform = ezPipeChannel_win;
+
 #endif
